requirements.txt
asgiref==3.2.3
Django==2.2.8
django-filter==2.2.0
djangorestframework==3.10.3
drf-nested-routers==0.91
Markdown==3.1.1
pytz==2019.3
sqlparse==0.3.0
